refactor: extract networking primitives to pyscv.net, add pyscv coverage
- Create src/pyscv/net.py with shared primitives: validate_url,
  safe_filename, resolve_url, atomic_download, run_gh
- Update download_artifacts.py to import from net.py
- Split tests: move networking tests to tests/test_net.py
- Add pyscv to pytest --cov and coverage.run.source_pkgs
- 100% branch coverage on all pyscv modules
